requirements.txt
discord.py
aiosqlite
Pillow
quart
python-dotenv
